refactor extract ServiceHelper from slobs 0.16.3
