+++ app/services/hotkeys.ts
@@ -3,8 +3,7 @@ import { ScenesService } from 'services/scenes';
 import { SourcesService } from 'services/sources';
 import { TransitionsService } from 'services/transitions';
 import { KeyListenerService } from 'services/key-listener';
-import { Inject } from 'services/core/injector';
-import { StatefulService, mutation, ServiceHelper } from 'services/core/stateful-service';
+import { StatefulService, mutation, ServiceHelper, Inject } from './core';
 import { defer } from 'lodash';
 import { $t } from 'services/i18n';
