+++ app/services/performance-monitor/performance-monitor.ts
@@ -1,6 +1,5 @@
-import { StatefulService, mutation, Service } from '../core/stateful-service';
+import { StatefulService, mutation, Service, Inject } from '../core';
 import * as obs from '../../../obs-api';
-import { Inject } from '../core/injector';
 import { NotificationsService, ENotificationType } from 'services/notifications';
 import { ServicesManager } from '../../services-manager';
 import { PerformanceService } from 'services/performance';
