+++ app/services/sources/source.ts
@@ -8,8 +8,7 @@ import {
   ISourceComparison,
   PROPERTIES_MANAGER_TYPES
 } from './index';
-import { mutation, ServiceHelper } from 'services/core/stateful-service';
-import { Inject } from 'services/core/injector';
+import { mutation, ServiceHelper, Inject } from '../core';
 import { ScenesService } from 'services/scenes';
 import { TObsFormData } from 'components/obs/inputs/ObsInput';
 import Utils from 'services/utils';
